feat: added permissions and routes

+++ src/controllers/auth.js
@@ -7,7 +7,7 @@
 const User = require('../models/user')
 const Token = require('../models/token')
 const passwordEncrypt = require('../helpers/passwordEncrypt')
-const jwt = require("jsonwebtoken")
+const jwt = require('jsonwebtoken')
 
 module.exports = {
 
@@ -48,37 +48,35 @@ module.exports = {
                     /* SIMPLE TOKEN */
 
                     /* JWT */
-                    
-                    const accessInfo = {          //* Kisa omurlu kritik olan bilgiler.
+
+                    const accessInfo = {
                         key: process.env.ACCESS_KEY,
-                        time: '30m',
+                        time: process.env?.ACCESS_EXP || '30m',
                         data: {
+                            _id: user.id,
                             id: user.id,
                             username: user.username,
                             email: user.email,
                             password: user.password,
                             isActive: user.isActive,
-                            isAdmin: user.isAdmin
-                        }
-
+                            isAdmin: user.isAdmin,
+                        },
                     }
-                    
-                    const refreshInfo = {      //* Uzun omurlu kritik olmayan bilgiler.
+
+                    const refreshInfo = {
                         key: process.env.REFRESH_KEY,
-                        time: '3d',
+                        time: process.env?.REFRESH_EXP || '30m',
                         data: {
                             id: user.id,
-                            password: user.password  // encrypted password
-                        }
+                            password: user.password // encrypted password
+                        },
                     }
 
-                    // jwt.sign(data, secret_key, { expiresIn: '30m' })
+                    // jwt.sign(access_data, access_key, { expiresIn: '30m' })
                     const accessToken = jwt.sign(accessInfo.data, accessInfo.key, { expiresIn: accessInfo.time })
                     const refreshToken = jwt.sign(refreshInfo.data, refreshInfo.key, { expiresIn: refreshInfo.time })
-                    
 
-
-                    /* JWT */ 
+                    /* JWT */
 
                     res.status(200).send({
                         error: false,
@@ -109,6 +107,51 @@ module.exports = {
 
     },
 
+    refresh: async (req, res) => {
+        /*
+            #swagger.tags = ["Authentication"]
+            #swagger.summary = "JWT: Refresh"
+            #swagger.description = 'Refresh token.'
+        */
+
+        const refreshToken = req.body?.bearer?.refresh
+
+        if (refreshToken) {
+
+            const refreshData = await jwt.verify(refreshToken, process.env.REFRESH_KEY)
+            // console.log(refreshData)
+
+            if (refreshData) {
+
+                const user = await User.findOne({ _id: refreshData.id })
+
+                if (user && user.password == refreshData.password) {
+
+                    res.status(200).send({
+                        error: false,
+                        bearer: {
+                            access: jwt.sign(user.toJSON(), process.env.ACCESS_KEY, { expiresIn: (process.env?.ACCESS_EXP || '30m') })
+                        }
+                    })
+
+                } else {
+
+                    res.errorStatusCode = 401
+                    throw new Error('Wrong id or password.')
+                }
+
+            } else {
+                res.errorStatusCode = 401
+                throw new Error('JWT refresh data is wrong.')
+            }
+
+        } else {
+            res.errorStatusCode = 401
+            throw new Error('Please enter bearer.refresh')
+        }
+
+    },
+
     logout: async (req, res) => {
         /*
             #swagger.tags = ["Authentication"]
@@ -118,13 +161,23 @@ module.exports = {
 
         const auth = req.headers?.authorization // Token ...tokenKey...
         const tokenKey = auth ? auth.split(' ') : null // ['Token', '...tokenKey...']
-        const result = await Token.deleteOne({ token: tokenKey[1] })
 
-        res.send({
-            error: false,
-            message: 'Token deleted. Logout was OK.',
-            result
-        })
+
+        if (tokenKey[0] == 'Token') {
+
+            const result = await Token.deleteOne({ token: tokenKey[1] })
+            res.send({
+                error: false,
+                message: 'Token deleted. Logout was OK.',
+                result
+            })
+
+        } else {
+
+            res.send({
+                error: false,
+                message: 'JWT: No need any process for logout.',
+            })
+        }
     }
-
 }
